% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crime}
\alias{crime}
\title{U.S. Statewide Crime data from the year 2003}
\format{
\subsection{\code{crime}}{

A data frame with 51 rows and 9 columns:
\describe{
\item{state}{The list of 50 states in the United States and the District of Columbia.}
\item{violent crime rate}{The annual number of murders, forcible rapes, robberies, and aggravated assaults per 100,000 people in the population.}
\item{murder rate}{The annual number of murders per 100,000 people in the population.}
\item{poverty}{Percentage of the residents with income below the poverty level.}
\item{high school}{Percentage of the adult residents who have at least a high school education.}
\item{college}{Percentage of the adult residents who have a college education.}
\item{single parent}{Percentage of families headed by a single parent.}
\item{unemployed}{Percentage of the adult residents who are unemployed.}
\item{metropolitan}{Percentage of the residents living in metropolitan areas.}
}
}
}
\source{
\url{https://img1.wsimg.com/blobby/go/bbca5dba-4947-4587-b40a-db346c01b1b3/downloads/us_statewide_crime.csv?ver=1709965708861}
}
\usage{
crime
}
\description{
This dataset is sourced from the 5th edition of "The Art and Science of Learning from Data" by Alan Agresti and Christine Franklin.
}
\keyword{datasets}
