% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUD-package.R
\docType{package}
\name{GUD}
\alias{GUD-package}
\alias{GUD}
\title{The 'GUD' package.}
\description{
This R package encompasses the probability density functions of three key distributions: the flexible Gumbel distribution, the double two-piece Student-t distribution, and the two-piece scale Student-t distribution, all belonging to the general unimodal distribution family, along with their corresponding sampling algorithms. Additionally, the package offers a function for Bayesian linear modal regression, leveraging these three distributions for model fitting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rh8liuqy/Bayesian_modal_regression}
}

}
\author{
\strong{Maintainer}: Qingyang Liu \email{qingyang@email.sc.edu} (\href{https://orcid.org/0000-0003-3265-6330}{ORCID})

Authors:
\itemize{
  \item Xianzheng Huang \email{huang@stat.sc.edu} (\href{https://orcid.org/0000-0001-7077-0869}{ORCID})
  \item Ray Bai \email{rbai@mailbox.sc.edu} (\href{https://orcid.org/0000-0002-7190-7844}{ORCID})
}

}
